{{cookiecutter.project_slug}}/requirements/base.txt
pytz==2019.1  # https://github.com/stub42/pytz
python-slugify==3.0.2  # https://github.com/un33k/python-slugify
Pillow==6.0.0  # https://github.com/python-pillow/Pillow
{%- if cookiecutter.use_compressor == "y" %}
rcssmin==1.0.6{% if cookiecutter.windows == 'y' and cookiecutter.use_docker == 'n' %} --install-option="--without-c-extensions"{% endif %}  # https://github.com/ndparker/rcssmin
{%- endif %}
argon2-cffi==19.1.0  # https://github.com/hynek/argon2_cffi
{%- if cookiecutter.use_whitenoise == 'y' %}
whitenoise==4.1.2  # https://github.com/evansd/whitenoise
{%- endif %}
redis==3.2.1  # https://github.com/antirez/redis
{%- if cookiecutter.use_celery == "y" %}
celery==4.3.0  # pyup: < 5.0  # https://github.com/celery/celery
django-celery-beat==1.5.0  # https://github.com/celery/django-celery-beat
{%- if cookiecutter.use_docker == 'y' %}
flower==0.9.3  # https://github.com/mher/flower
{%- endif %}
{%- endif %}

# Django
# ------------------------------------------------------------------------------
django==2.2.2  # pyup: < 3.0  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==3.1.2  # https://github.com/jazzband/django-model-utils
django-allauth==0.39.1  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.7.2  # https://github.com/django-crispy-forms/django-crispy-forms
{%- if cookiecutter.use_compressor == "y" %}
django-compressor==2.3  # https://github.com/django-compressor/django-compressor
{%- endif %}
django-redis==4.10.0  # https://github.com/niwinz/django-redis

# Django REST Framework
djangorestframework==3.9.4  # https://github.com/encode/django-rest-framework
coreapi==2.3.3  # https://github.com/core-api/python-client


# pai

# CORS
django-cors-headers==3.0.2

# django-sso
{%- if cookiecutter.use_django_sso_app_profiles == 'y' %}
django-sso-app
{%- endif %}
{%- if cookiecutter.use_django_rest_framework_kong_consumers == 'y' %}
django-rest-framework-kong-consumers
{%- endif %}
